Change the namespace for clarity (Drupal\entity\Query -> Drupal\entity\QueryAccess)
