+++ src/QueryAccess/QueryAccessHandler.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\entity\Query;
+namespace Drupal\entity\QueryAccess;
 
 use Drupal\Core\Session\AccountInterface;
