+++ src/QueryAccess/UncacheableQueryAccessHandler.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\entity\Query;
+namespace Drupal\entity\QueryAccess;
 
 /**
  * Controls query access based on the uncacheable entity permissions.
